Please enter the commit message for your changes. Lines starting

+++ udc_inputs.py
@@ -10,7 +10,7 @@ def get_feature_columns(mode):
   feature_columns.append(tf.contrib.layers.real_valued_column(
       column_name="context_len", dimension=1, dtype=tf.int64))
   feature_columns.append(tf.contrib.layers.real_valued_column(
-      column_name="context_token_len_avg", dimension=1, dtype=tf.float64))
+      column_name="context_token_len_avg", dimension=1, dtype=tf.float32))
   feature_columns.append(tf.contrib.layers.real_valued_column(
       column_name="context_nums", dimension=1, dtype=tf.int64))  
 
@@ -19,7 +19,7 @@ def get_feature_columns(mode):
   feature_columns.append(tf.contrib.layers.real_valued_column(
       column_name="utterance_len", dimension=1, dtype=tf.int64))
   feature_columns.append(tf.contrib.layers.real_valued_column(
-      column_name="utterance_token_len_avg", dimension=1, dtype=tf.float64))
+      column_name="utterance_token_len_avg", dimension=1, dtype=tf.float32))
   feature_columns.append(tf.contrib.layers.real_valued_column(
       column_name="utterance_nums", dimension=1, dtype=tf.int64))  
   if mode == tf.contrib.learn.ModeKeys.TRAIN:
